+++ CHANGELOG.md
@@ -4,11 +4,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
+# v2.11.0 
+
+### Added 
+
+- `loglink` command, returns the log link for the current thread.
+
+# v2.10.2
+- Your logs now track and show edited messages.
+
+# v2.10.1
 
 ### Changed
-- Improve format of thread message embeds. Slightly cleaner and simpler now.
+- Use reply author's top role for the mod tag by default.
+
+# v2.10.0
+
+### Added
+- `anonreply` command to anonymously reply to the recipient. 
+The username of the anonymous user defaults to the `mod_tag` (the footer text of a mod reply). The avatar defaults the guild icon url. However you can change both of these via the `anon_username`, `anon_avatar_url` and `anon_tag` config variables. 
+
+### Changed
+Your bot now logs all messages sent in a thread channel, including discussions that take place. You can now toggle to view them in the log viewer app.
+
+# v2.9.4
+Fixed a small bug due to a typo.
+
+# v2.9.3
+Forgot to enable custom embed colors.
+
+### Added
+- Ability to set a custom `mod_tag` (the text in the footer of the mod reply embed, which by default says "Moderator")
+
+# v2.9.2
+
+### Changed
+- Improve format of thread info embed. Slightly cleaner and simpler now.
 - All commands are now blurple instead of green.
+
+### Fixed
+- Bug where the close command wouldn't work if you didnt configure a log channel. 
 
 ### Added
 - Ability to set your own custom `mod_color` and `recipient_color` for the thread message embeds.